+++ src/mods/scala/domains/terrestrial/Output.scala
@@ -11,8 +11,9 @@ object Output extends App with DracoType {
   override lazy val typeDefinition: TypeDefinition = TypeLoader.loadType(TypeName ("Output", _namePackage = Seq ("domains", "terrestrial")))
   lazy val dracoType: Type[Output] = Type[Output] (typeDefinition)
 
-  def actorType(consumer: ActorRef[draco.format.json.JSON]): ActorType = new Actor[domains.world.World] {
+  def actorType(_consumer: => ActorRef[draco.format.json.JSON]): ActorType = new Actor[domains.world.World] {
     override lazy val typeDefinition: TypeDefinition = Output.typeDefinition
+    lazy val consumer: ActorRef[draco.format.json.JSON] = _consumer
 
     override def receive(ctx: TypedActorContext[domains.world.World], msg: domains.world.World): Behavior[domains.world.World] = {
       msg match {